Feat(#54): relocate s3Provider
