+++ src/main/java/com/numberone/backend/domain/shelter/service/ShelterService.java
@@ -5,7 +5,7 @@ import com.numberone.backend.domain.shelter.dto.response.*;
 import com.numberone.backend.domain.shelter.repository.ShelterRepository;
 import com.numberone.backend.domain.shelter.util.ShelterType;
 import com.numberone.backend.exception.notfound.NotFoundShelterException;
-import com.numberone.backend.support.S3Provider;
+import com.numberone.backend.support.s3.S3Provider;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
